requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
pandas
pydantic
python-dotenv
streamlit
requests